requirements.txt
# PyTorch
torch>=0.4.1
# progress bars in model download and training scripts
tqdm
# Accessing files from S3 directly.
boto3
# Used for downloading models over HTTP
requests
numpy==1.14.3
joblib==0.13.2
paddlepaddle